+++ src/test/kotlin/stringcalculator/StringAddCalculatorTest.kt
@@ -1,3 +1,5 @@
+package stringcalculator
+
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatExceptionOfType
 import org.junit.jupiter.api.BeforeEach
@@ -6,7 +8,6 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.NullAndEmptySource
 import org.junit.jupiter.params.provider.ValueSource
-import stringcalculator.StringAddCalculator
 
 class StringAddCalculatorTest {
 
@@ -21,41 +22,41 @@ class StringAddCalculatorTest {
     @ParameterizedTest
     @NullAndEmptySource
     fun emptyOrNull(text: String?) {
-        assertThat(calculator.add(text)).isZero()
+        assertThat(StringAddCalculator.add(text)).isZero()
     }
 
     @DisplayName(value = "숫자 하나를 문자열로 입력할 경우 해당 숫자를 반환한다.")
     @ParameterizedTest
     @ValueSource(strings = ["1"])
     fun oneNumber(text: String) {
-        assertThat(calculator.add(text)).isSameAs(Integer.parseInt(text))
+        assertThat(StringAddCalculator.add(text)).isSameAs(Integer.parseInt(text))
     }
 
     @DisplayName(value = "숫자 두개를 쉼표(,) 구분자로 입력할 경우 두 숫자의 합을 반환한다.")
     @ParameterizedTest
     @ValueSource(strings = ["1,2"])
     fun twoNumbers(text: String) {
-        assertThat(calculator.add(text)).isSameAs(3)
+        assertThat(StringAddCalculator.add(text)).isSameAs(3)
     }
 
     @DisplayName(value = "구분자를 쉼표(,) 이외에 콜론(:)을 사용할 수 있다.")
     @ParameterizedTest
     @ValueSource(strings = ["1,2:3"])
     fun colons(text: String) {
-        assertThat(calculator.add(text)).isSameAs(6)
+        assertThat(StringAddCalculator.add(text)).isSameAs(6)
     }
 
     @DisplayName(value = "//와 \\n 문자 사이에 커스텀 구분자를 지정할 수 있다.")
     @ParameterizedTest
     @ValueSource(strings = ["//;\n1;2;3"])
     fun customDelimiter(text: String) {
-        assertThat(calculator.add(text)).isSameAs(6)
+        assertThat(StringAddCalculator.add(text)).isSameAs(6)
     }
 
     @DisplayName(value = "문자열 계산기에 음수를 전달하는 경우 RuntimeException 예외 처리를 한다.")
     @Test
     fun negative() {
         assertThatExceptionOfType(RuntimeException::class.java)
-            .isThrownBy { calculator.add("-1") }
+            .isThrownBy { StringAddCalculator.add("-1") }
     }
 }
